refactor: split led-diode-string.test.ts into one test per file
Split the combined test file into 8 separate files (led-diode-string1.test.ts
through led-diode-string8.test.ts) to comply with the repo rule that each
*.test.ts file should have at most one test() call.
